+++ Flow/Changes.md
@@ -4,9 +4,12 @@
 ### Breaking Changes
 - `.divide(String)` in **Regex** now returns `Vector[Either[String, String]]` instead of `Vector[String]`
   - The previous implementation is now named `.separate(String)`
+- **Loop** is now an abstract class instead of a trait. This is because the trait contains attributes.
 ### New Methods
 - **Lazy** (object)
   - Added `.wrap(...)`, which calls `LazyWrapper.apply(...)`
+- **Loop**
+  - Added `.startAsyncAfter(FiniteDuration)`, a delayed version of `.startAsync()`
 - **ResettableLazy**
   - Added `.filter(...)` and `.filterNot(...)`, which are kind of conditional resets